# cmd/api/docs/swagger.yaml
basePath: /v1
definitions:
  cerrors.IError:
    type: object
...
  docstring.Typedef:
    properties:
      args:
        items:
          $ref: '#/definitions/docstring.TypedefArg'
        type: array
      name:
        type: string
      type:
        type: string
    type: object
  docstring.TypedefArg:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  gin.H:
    additionalProperties: true
    type: object
...
  handlers.Contract:
    properties:
      address:
        type: string
      alias:
        type: string
      annotations:
        items:
          type: string
        type: array
      balance:
        type: integer
      delegate:
        type: string
      delegate_alias:
        type: string
      entrypoints:
        items:
          type: string
        type: array
      fail_strings:
        items:
          type: string
        type: array
      fingerprint:
        $ref: '#/definitions/models.Fingerprint'
        type: object
      found_by:
        type: string
      hardcoded:
        items:
          type: string
        type: array
      hash:
        type: string
      id:
        type: string
      language:
        type: string
      last_action:
        $ref: '#/definitions/models.BCDTime'
        type: object
      level:
        type: integer
      manager:
        type: string
      migrations_count:
        type: integer
      network:
        type: string
      profile:
        $ref: '#/definitions/handlers.ProfileInfo'
        type: object
      project_id:
        type: string
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      timestamp:
        type: string
      total_withdrawn:
        type: integer
      tx_count:
        type: integer
    type: object
  handlers.EntrypointSchema:
    properties:
      bin_path:
        type: string
      default_model:
        $ref: '#/definitions/jsonschema.DefaultModel'
        type: object
      name:
        type: string
      schema:
        $ref: '#/definitions/jsonschema.Schema'
        type: object
      typedef:
        items:
          $ref: '#/definitions/docstring.Typedef'
        type: array
    type: object
  handlers.Error:
    properties:
      message:
        example: text
        type: string
    type: object
...
  handlers.Migration:
    properties:
      hash:
        type: string
      kind:
        type: string
      level:
        type: integer
      prev_protocol:
        type: string
      protocol:
        type: string
      timestamp:
        type: string
    type: object
...
  handlers.Operation:
    properties:
      amount:
        type: integer
      balance:
        type: integer
      balance_updates:
        items:
          $ref: '#/definitions/models.BalanceUpdate'
        type: array
      burned:
        type: integer
      counter:
        type: integer
      delegate:
        type: string
      destination:
        type: string
      destination_alias:
        type: string
      entrypoint:
        type: string
      errors:
        items:
          $ref: '#/definitions/cerrors.IError'
        type: array
      fee:
        type: integer
      gas_limit:
        type: integer
      hash:
        type: string
      id:
        type: string
      internal:
        type: boolean
      kind:
        type: string
      level:
        type: integer
      manager_pubkey:
        type: string
      mempool:
        type: boolean
      network:
        type: string
      parameters:
        type: object
      protocol:
        type: string
      public_key:
        type: string
      result:
        $ref: '#/definitions/models.OperationResult'
        type: object
      source:
        type: string
      source_alias:
        type: string
      status:
        type: string
      storage_diff:
        type: object
      storage_limit:
        type: integer
      timestamp:
        type: string
    type: object
...
  handlers.ProfileInfo:
    properties:
      subscribed:
        type: boolean
    type: object
...
  jsonschema.DefaultModel:
    additionalProperties: true
    type: object
  jsonschema.Schema:
    additionalProperties: true
    type: object
  models.BCDTime:
    type: object
  models.BalanceUpdate:
    properties:
      category:
        type: string
      change:
        type: integer
      contract:
        type: string
      cycle:
        type: integer
      delegate:
        type: string
      kind:
        type: string
    type: object
...
  models.Fingerprint:
    properties:
      code:
        type: string
      parameter:
        type: string
      storage:
        type: string
    type: object
  models.OperationResult:
    properties:
      allocated_destination_contract:
        type: boolean
      balance_updates:
        items:
          $ref: '#/definitions/models.BalanceUpdate'
        type: array
      consumed_gas:
        type: integer
      paid_storage_size_diff:
        type: integer
      storage_size:
        type: integer
    type: object
...
host: api.better-call.dev
info:
  contact: {}
  description: This is API description for Better Call Dev service.
  license: {}
  title: Better Call Dev API
  version: "1.0"
paths:
  /contract/{network}/{address}:
    get:
      consumes:
      - application/json
      description: Get full contract info
      operationId: get-contract
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Contract'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract info
      tags:
      - contract
...
  /contract/{network}/{address}/entrypoints:
    get:
      consumes:
      - application/json
      description: Get contract entrypoints
      operationId: get-contract-entrypoints
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.EntrypointSchema'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract entrypoints
      tags:
      - contract
  /contract/{network}/{address}/entrypoints/data:
    post:
      consumes:
      - application/json
      description: Get entrypoint data from schema object
      operationId: get-contract-entrypoints-data
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      - description: Entrypoint binary path
        in: body
        name: bin_path
        required: true
        schema:
          type: string
      - description: Entrypoint`s arguments data
        in: body
        name: data
        required: true
        schema:
          type: object
      - description: Return format (michelson | empty if micheline)
        in: body
        name: format
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get entrypoint data from schema object
      tags:
      - contract
...
  /contract/{network}/{address}/mempool:
    get:
      consumes:
      - application/json
      description: Get contract mempool operations
      operationId: get-contract-mempool
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.Operation'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract mempool operations
      tags:
      - contract
  /contract/{network}/{address}/migrations:
    get:
      consumes:
      - application/json
      description: Get contract migrations
      operationId: get-contract-migrations
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.Migration'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract migrations
      tags:
      - contract
...
swagger: "2.0"